+++ src/PhpWord/Element/Field.php
@@ -91,10 +91,10 @@ class Field extends AbstractElement
             ],
             'options' => ['Path', 'PreserveFormat'],
         ],
-        'REF' => array(
-            'properties' => array('name' => ''),
-            'options'    => array('f', 'h', 'n', 'p', 'r', 't', 'w'),
-        ),
+        'REF' => [
+            'properties' => ['name' => ''],
+            'options' => ['f', 'h', 'n', 'p', 'r', 't', 'w'],
+        ],
     ];
 
     /**
